IntegrationBeeApp/requirements.txt
anyascii==0.3.2
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-libsass==0.9
django-modelcluster==6.3
django-permissionedforms==0.1
django-sass-processor==1.4.1
django-taggit==5.0.0
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
executing==2.1.0
filetype==1.2.0
html5lib==1.1
idna==3.10
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
l18n==2021.3
laces==0.1.1
libsass==0.23.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mysqlclient==2.2.5
openpyxl==3.1.5
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
pillow_heif==0.20.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pytz==2024.2
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.3
telepath==0.3.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
wagtail==6.2.2
wagtail-font-awesome-svg==1.0.1
wagtailgeowidget==8.1.1
wagtailgmaps==1.0.1
wagtailmath==1.3.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.9.0

pillow==10.4.0

